Bugfix/convert count values to usize (#48)
* convert UnmanagedQueue.count to usize

* convert ManagedQueue.count to usize

+++ src/managed_queue.zig
@@ -27,7 +27,7 @@ pub fn ManagedQueue(comptime T: type) type {
         allocator: std.mem.Allocator,
         head: ?*Node,
         tail: ?*Node,
-        count: u32,
+        count: usize,
 
         pub fn init(allocator: std.mem.Allocator) Self {
             return Self{
